Reorganize CFG utilities
